+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Add `MemberChecks.mustBeProtected(...)` and `MemberChecks.mustNotBeProtected(...)` helper methods.
+
 ### [v0.50](https://github.com/realityforge/proton/tree/v0.50) (2020-05-21) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.49...v0.50)
 
 Changes in this release: